improve snippet formatting
removed type annotations from array creation

replaced another "double" with "float"

normalized formatting

+++ samples/snippets/fsharp/VS_Snippets_CLR_System/system.Random/fs/doublerange1.fs
@@ -10,7 +10,7 @@ let rnd = Random()
 
 // Generate 20 million random integers.
 let count =
-    Array.init 20000000 (fun _ -> rnd.NextDouble() * (double Int64.MaxValue) )
+    Array.init 20000000 (fun _ -> rnd.NextDouble() * (float Int64.MaxValue) )
     |> Array.countBy (fun x -> x / (float ONE_TENTH) |> int ) // Categorize into 10 groups and count them.
     |> Array.map snd
